+++ lib/dart_esr.dart
@@ -7,6 +7,8 @@ export 'src/eosio_signing_request.dart';
 
 export 'src/utils/esr_constant.dart';
 
-export 'src/models/identity.dart';
-export 'src/models/transaction.dart';
 export 'src/models/action.dart';
+export 'src/models/authorization.dart';
+export 'src/models/identity.dart';
+export 'src/models/signing_request.dart';
+export 'src/models/transaction.dart';
