+++ lib/src/models/identity.dart
@@ -1,14 +1,16 @@
 import 'dart:typed_data';
+
+import 'package:dart_esr/src/models/authorization.dart';
 import 'package:json_annotation/json_annotation.dart';
 
-import 'package:eosdart/eosdart.dart';
+import 'package:eosdart/eosdart.dart' as ser;
 
 part 'identity.g.dart';
 
 @JsonSerializable(explicitToJson: true)
 class Identity {
   @JsonKey(name: 'permission')
-  IdentityPermission identityPermission;
+  Authorization authorization;
 
   Identity();
 
@@ -20,28 +22,9 @@ class Identity {
   @override
   String toString() => this.toJson().toString();
 
-  Uint8List toBinary(Type type) {
-    var buffer = SerialBuffer(Uint8List(0));
+  Uint8List toBinary(ser.Type type) {
+    var buffer = ser.SerialBuffer(Uint8List(0));
     type.serialize(type, buffer, this.toJson());
     return buffer.asUint8List();
   }
-}
-
-@JsonSerializable(explicitToJson: true)
-class IdentityPermission {
-  @JsonKey(name: 'actor')
-  String actor;
-
-  @JsonKey(name: 'permission')
-  String permission;
-
-  IdentityPermission();
-
-  factory IdentityPermission.fromJson(Map<String, dynamic> json) =>
-      _$IdentityPermissionFromJson(json);
-
-  Map<String, dynamic> toJson() => _$IdentityPermissionToJson(this);
-
-  @override
-  String toString() => this.toJson().toString();
 }
