+++ lib/src/models/transaction.dart
@@ -1,88 +1,10 @@
 import 'dart:typed_data';
-import 'package:dart_esr/src/utils/conversion_helper.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 import 'package:dart_esr/src/models/action.dart';
 import 'package:eosdart/eosdart.dart' as ser;
 
 part 'transaction.g.dart';
-
-@JsonSerializable()
-class TransactionBlock with ConversionHelper {
-  @JsonKey(name: 'id')
-  String id;
-
-  @JsonKey(name: 'trx')
-  TransactionWithReceipt trx;
-
-  @JsonKey(name: 'block_time')
-  DateTime blockTime;
-
-  @JsonKey(name: 'block_num', fromJson: ConversionHelper.getIntFromJson)
-  int blockNum;
-
-  @JsonKey(
-      name: 'last_irreversible_block',
-      fromJson: ConversionHelper.getIntFromJson)
-  int lastIrreversibleBlock;
-
-  @JsonKey(name: 'traces')
-  List<ActionWithReceipt> traces;
-
-  TransactionBlock();
-
-  factory TransactionBlock.fromJson(Map<String, dynamic> json) =>
-      _$TransactionBlockFromJson(json);
-
-  Map<String, dynamic> toJson() => _$TransactionBlockToJson(this);
-
-  @override
-  String toString() => this.toJson().toString();
-}
-
-@JsonSerializable()
-class TransactionWithReceipt {
-  @JsonKey(name: 'receipt')
-  TransactionReceipt receipt;
-
-  @JsonKey(name: 'trx')
-  Transaction transaction;
-
-  TransactionWithReceipt();
-
-  factory TransactionWithReceipt.fromJson(Map<String, dynamic> json) =>
-      _$TransactionWithReceiptFromJson(json);
-
-  Map<String, dynamic> toJson() => _$TransactionWithReceiptToJson(this);
-
-  @override
-  String toString() => this.toJson().toString();
-}
-
-@JsonSerializable()
-class TransactionReceipt with ConversionHelper {
-  @JsonKey(name: 'status')
-  String status;
-
-  @JsonKey(name: 'cpu_usage_us', fromJson: ConversionHelper.getIntFromJson)
-  int cpuUsageUs;
-
-  @JsonKey(name: 'net_usage_words', fromJson: ConversionHelper.getIntFromJson)
-  int netUsageWords;
-
-  @JsonKey(name: 'trx')
-  Object trx;
-
-  TransactionReceipt();
-
-  factory TransactionReceipt.fromJson(Map<String, dynamic> json) =>
-      _$TransactionReceiptFromJson(json);
-
-  Map<String, dynamic> toJson() => _$TransactionReceiptToJson(this);
-
-  @override
-  String toString() => this.toJson().toString();
-}
 
 @JsonSerializable(explicitToJson: true)
 class Transaction {
@@ -134,63 +56,4 @@ class Transaction {
     transactionType.serialize(transactionType, buffer, this.toJson());
     return buffer.asUint8List();
   }
-}
-
-@JsonSerializable()
-class TransactionCommitted {
-  @JsonKey(name: 'transaction_id')
-  String id;
-
-  @JsonKey(name: 'processed')
-  TransactionProcessed processed;
-
-  TransactionCommitted();
-
-  factory TransactionCommitted.fromJson(Map<String, dynamic> json) =>
-      _$TransactionCommittedFromJson(json);
-
-  Map<String, dynamic> toJson() => _$TransactionCommittedToJson(this);
-
-  @override
-  String toString() => this.toJson().toString();
-}
-
-@JsonSerializable()
-class TransactionProcessed with ConversionHelper {
-  @JsonKey(name: 'id')
-  String id;
-
-  @JsonKey(name: 'block_num', fromJson: ConversionHelper.getIntFromJson)
-  int blockNum;
-
-  @JsonKey(name: 'block_time')
-  DateTime blockTime;
-
-  @JsonKey(name: 'producer_block_id', fromJson: ConversionHelper.getIntFromJson)
-  int producerBlockId;
-
-  @JsonKey(name: 'receipt')
-  TransactionReceipt receipt;
-
-  @JsonKey(name: 'elapsed', fromJson: ConversionHelper.getIntFromJson)
-  int elapsed;
-
-  @JsonKey(name: 'net_usage', fromJson: ConversionHelper.getIntFromJson)
-  int netUsage;
-
-  @JsonKey(name: 'scheduled')
-  bool scheduled;
-
-  @JsonKey(name: 'action_traces')
-  List<ActionWithReceipt> actionTraces;
-
-  TransactionProcessed();
-
-  factory TransactionProcessed.fromJson(Map<String, dynamic> json) =>
-      _$TransactionProcessedFromJson(json);
-
-  Map<String, dynamic> toJson() => _$TransactionProcessedToJson(this);
-
-  @override
-  String toString() => this.toJson().toString();
 }
